feat: move out clone opts
